% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateInteractionsToGInteractions.R
\name{CandidateInteractionsToGInteractions}
\alias{CandidateInteractionsToGInteractions}
\title{CandidateInteractionsToGInteractions
transform Chicdiff CandidateInteractions (data frame) To GInteractions object using rmap}
\usage{
CandidateInteractionsToGInteractions(
  candidateInteractionsDataFrame,
  pathToRmap
)
}
\arguments{
\item{chicDiffResultsDataFrame}{("data.table" or "data.frame") Results dataset from Chicdiff pipeline}

\item{path}{to rmap file that was used upstream in chicago/chicdiff. Rmap file should have four columns: V1:V4 ("chr", "start", "end", "RF_n")}
}
\value{
GInteractions object with leftover columns forming mcols
}
\description{
CandidateInteractionsToGInteractions
transform Chicdiff CandidateInteractions (data frame) To GInteractions object using rmap
}
