% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicDiffToGI.R
\name{chicDiffToGI}
\alias{chicDiffToGI}
\title{chicDiffToGI
converts ChicDiff resultsMatrix to a GInteractions object}
\usage{
chicDiffToGI(ChicDiffResultsDataFrame)
}
\arguments{
\item{ChicDiffResultsDataFrame}{ChicDiff results data frame}
}
\value{
GInteractions object with baitchr, baitstart, baitend forming anchorOne and OEchr, OEstart, OEend forming anchorTwo. The remaining columns are stored in mcols().
}
\description{
chicDiffToGI
converts ChicDiff resultsMatrix to a GInteractions object
}
