% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countChIPPeaksInPromoters.R
\name{countChIPPeaksInPromoters}
\alias{countChIPPeaksInPromoters}
\title{countChIPPeaksInPromoters
count ChIP peaks in promoters of given coordinates.}
\usage{
countChIPPeaksInPromoters(pathToSnakePipesDir, GRangesObject, titleString)
}
\arguments{
\item{pathToSnakePipesDir}{(string) path to snakePipes Chip output directory (for example CSAW_Gencode_ATAC or CSAW_MACS2_ATAC)}

\item{GRangesObject}{the genomic ranges (such as expressed gene promoters coordinates) around which to count the ChIP peaks}

\item{titleString}{the title for title}
}
\value{
ggplot barplot object with overlap counts where x axis has ChIP peaks class, y axis contains the promoter count and colors depict the gene regulation group.
}
\description{
countChIPPeaksInPromoters
count ChIP peaks in promoters of given coordinates.
}
