% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFeatureSet.R
\name{countFeatureSet}
\alias{countFeatureSet}
\title{countFeatureSet
counts overlaps between \code{GRangesObject} and three cathegories of peaks from snakePipes results directory specified by \code{ChipNamesVector}
The three cathegories of peaks are: snakePipesDir/DiffBind_all.tsv (peaks_all), snakePipesDir/DiffBind_UP.FilteredResults.tsv (peaks_up), snakePipesDir/DiffBind_DOWN.FilteredResults.tsv (peaks_down) <- make correct the filenames}
\usage{
countFeatureSet(GRangesObject, ChipNamesVector)
}
\value{
a data.frame with nrows of GRangesObject and ncols of 3*length(listOfChIPs) with counts of overlaps
}
\description{
countFeatureSet
counts overlaps between \code{GRangesObject} and three cathegories of peaks from snakePipes results directory specified by \code{ChipNamesVector}
The three cathegories of peaks are: snakePipesDir/DiffBind_all.tsv (peaks_all), snakePipesDir/DiffBind_UP.FilteredResults.tsv (peaks_up), snakePipesDir/DiffBind_DOWN.FilteredResults.tsv (peaks_down) <- make correct the filenames
}
