% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadsInGranges.R
\name{countReadsInGranges}
\alias{countReadsInGranges}
\title{countReadsInGranges counts reads in aligned BAM files for regions stored in GRanges by using Rsubread::Featurecounts.}
\usage{
countReadsInGranges(
  GRangesObject,
  snakePipesDirectoryString,
  feature,
  nthreads
)
}
\arguments{
\item{GRangesObject}{GRanges object for which featyres should be counted}

\item{snakePipesDirectoryString}{The full path to a snakePipes directory with filtered BAM files}

\item{feature}{of ChIPs - the name of ChIP to find the correct files in snakePipes BAM files.}

\item{nthreads}{number of threads}
}
\value{
the output of Rsubread::featureCounts() function (seee ?FeatureCounts 'Value')
}
\description{
countReadsInGranges counts reads in aligned BAM files for regions stored in GRanges by using Rsubread::Featurecounts.
}
