% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGOTerms.R
\name{findGOTerms}
\alias{findGOTerms}
\title{findGOTerms
Find a list of GO terms based on keywords based on constraints "A AND B"
Finds a list of GO terms based on two keywords in the \code{MouseMine} using \code{InterMineR} library}
\usage{
findGOTerms(KeywordsVector)
}
\arguments{
\item{KeywordsVector}{a character string vector of length 2}
}
\description{
findGOTerms
Find a list of GO terms based on keywords based on constraints "A AND B"
Finds a list of GO terms based on two keywords in the \code{MouseMine} using \code{InterMineR} library
}
\examples{
findGOTerms(c("histone", "H3-K4"))
}
