% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBed6AsGRanges.R
\name{loadBed6AsGRanges}
\alias{loadBed6AsGRanges}
\title{load snakePipes DiffBinding_allregions.bed BED6 file as GRanges object, giving appropriate column names
note that BED is a 0-based format, while GRanges is 1-based format
This means that the start coordinate should be converted to +1 when converting from BED to GRanges}
\usage{
loadBed6AsGRanges(pathToFile)
}
\arguments{
\item{pathToFile}{(string) an absolute or relative path to a bed file}
}
\description{
load snakePipes DiffBinding_allregions.bed BED6 file as GRanges object, giving appropriate column names
note that BED is a 0-based format, while GRanges is 1-based format
This means that the start coordinate should be converted to +1 when converting from BED to GRanges
}
