% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFilteredResultsAsGRanges.R
\name{loadFilteredResultsAsGRanges}
\alias{loadFilteredResultsAsGRanges}
\title{loads snakePipes Filtered.results.*.bed file as GRanges object, giving appropriate column names}
\usage{
loadFilteredResultsAsGRanges(pathToFile)
}
\arguments{
\item{pathToFile}{(string) an absolute or relative path to a file}
}
\description{
loads snakePipes Filtered.results.*.bed file as GRanges object, giving appropriate column names
}
