% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFeaturesList.R
\name{makeFeaturesList}
\alias{makeFeaturesList}
\title{creates a features list from snakePipes ChIP pipeline}
\usage{
makeFeaturesList(snakePipesOutputDir)
}
\arguments{
\item{snakePipesOutputDir}{(string) name of snakePipes ChIP pipeline output, for example '$PWD/DNA-mapping/CSAW_MACS2_sampleSheetName'}
}
\value{
GRangesList object with list items named as \code{peaks_all}, \code{peaks_up}, \code{peaks_down}
}
\description{
creates a features list from snakePipes ChIP pipeline
}
