% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureOverlaps.R
\name{plotFeatureOverlaps}
\alias{plotFeatureOverlaps}
\title{plot the overlaps with features in a particular way}
\usage{
plotFeatureOverlaps(long_tibble)
}
\arguments{
\item{long_tibble}{a 'long' format tibble with columns names \code{cathegory}(factor), \code{nonzero_features_count}(numeric), \code{Status}(factor)}
}
\value{
ggplot object
@export
}
\description{
a wrapper function to plot the overlap counts in a particular way
}
