% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAllAfterDot.R
\name{removeAllAfterDot}
\alias{removeAllAfterDot}
\title{removeAllAfterDot
a convenience function for extracting gene stable IDs from ensembl ID}
\usage{
removeAllAfterDot(x)
}
\arguments{
\item{x}{(string) Ensembl ID (e.g., ENMUSG123456.78)}
}
\value{
(string) with numbers after dot removed (e.g., ENMUSG123456)
}
\description{
removeAllAfterDot
a convenience function for extracting gene stable IDs from ensembl ID
}
