% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeOverlapsWithFeatureList.R
\name{summarizeOverlapsWithFeatureList}
\alias{summarizeOverlapsWithFeatureList}
\title{Summarize overlaps with a feature list}
\usage{
summarizeOverlapsWithFeatureList(GRanges_query, GRangesList_subject)
}
\arguments{
\item{GRanges_query}{GRanges object that will be used as a query}

\item{GRangesList_subject}{GRangesList object that will be used as a subject}
}
\value{
GRanges_query with overlaps from each item in GRangesList_object as a column in mcols()
}
\description{
counts overlaps between a GRanges object (query) in each item of GRangesList (many subjects)
outputs the count of overlaps for each GRanges element as a data frame
}
\examples{

# make a features list from snakePipes chip-seq pipeline output
featuresList = makeFeaturesList("/scratch/adrija/02_MLL/data/chip-seq/DNA-mapping/CSAW_MACS2_CEBPa")

# make a GRanges query from ATAC peaks
ATAC_peaks = omicsWrappers::loadBed6AsGRanges("/localhome/bric/qlr900/analysis/02_MLL/data/atac/DNA-mapping/CSAW_MACS2_ATAC/DiffBinding_allregions.bed")

summarize overlaps
summarizeOverlapsWithFeatureList(ATAC_peaks, featuresList)
}
