% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGRanges2BED.R
\name{writeGRanges2BED}
\alias{writeGRanges2BED}
\title{write GRanges object to BED6 format, converting 1-based GRanges coordinates to 0-based BED6 coordinates
this function is borrowed from Devon Ryan's answer at BioStars: https://www.biostars.org/p/89341/}
\usage{
writeGRanges2BED(gr, fileNameString)
}
\arguments{
\item{gr}{GRanges object}

\item{fileNameString}{a character string describing psth to the output BED file name}
}
\description{
write GRanges object to BED6 format, converting 1-based GRanges coordinates to 0-based BED6 coordinates
this function is borrowed from Devon Ryan's answer at BioStars: https://www.biostars.org/p/89341/
}
