% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBed6AsGRanges.R
\name{loadBed6AsGRanges}
\alias{loadBed6AsGRanges}
\title{loads snakePipes DiffBinding_allregions.bed BED6 file as GRanges object, giving appropriate column names}
\usage{
loadBed6AsGRanges(pathToFile)
}
\arguments{
\item{pathToFile}{(string) an absolute or relative path to a bed file}
}
\description{
loads snakePipes DiffBinding_allregions.bed BED6 file as GRanges object, giving appropriate column names
}
