% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFeaturesList.R
\name{makeFeaturesList}
\alias{makeFeaturesList}
\title{create a features list from snakePipes ChIP pipeline}
\usage{
makeFeaturesList(snakePipesOutputDir)
}
\arguments{
\item{snakePipesOutputDir}{(string) name of snakePipes ChIP pipeline output, for example '$PWD/DNA-mapping/CSAW_MACS2_sampleSheetName'}
}
\value{
GRangesList object with list items named as \code{peaks_all}, \code{peaks_up}, \code{peaks_down}
}
\description{
create a features list from snakePipes ChIP pipeline
}
\examples{
   snakeDir = "/localhome/bric/qlr900/analysis/02_MLL/data/chip-seq/DNA-mapping/CSAW2_MACS2_H3K27ac"
   makeFeaturesList(snakeDir)
}
